requirements.txt
dill==0.3.4
scikit-spatial==6.2.0
h5py==3.7.0
keras==2.10.0
matplotlib==3.3.2
opencv_python_headless==4.5.5.64
pandas==1.2.3
Pillow==9.2.0
pyotf==0.0.2
PyYAML==6.0
scikit_image==0.19.1
scikit_learn==1.1.2
scipy==1.8.0
seaborn==0.11.2
scikit-image
tensorflow==2.10.0
tensorflow_gpu==2.10.0
tifffile==2021.4.8
tqdm==4.50.2
pytest
numpy==1.23.4
